backend/requirements.txt
# 血浆游离血红蛋白检测系统 - 后端依赖包（完整版）
# 作者: 哈雷酱大小姐 (￣▽￣)／

# ============ FastAPI 核心依赖 ============
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# ============ 数据库相关 ============
sqlalchemy==2.0.25
aiomysql==0.3.2  # MySQL异步驱动
pymysql==1.1.2  # MySQL同步驱动（aiomysql依赖）
alembic==1.13.1
# aiosqlite==0.19.0  # SQLite驱动（如果需要SQLite，取消注释）
# asyncpg==0.29.0  # PostgreSQL驱动（如果需要PostgreSQL，取消注释）
# psycopg2-binary==2.9.9  # PostgreSQL适配器（如果需要PostgreSQL，取消注释）

# ============ 机器学习 ============
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.3
joblib==1.3.2

# ============ 数据处理 ============
openpyxl==3.1.2
pandas==2.2.0

# ============ 用户认证 ============
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic-settings==2.1.0

# ============ PDF生成 ============
reportlab==4.0.7
fpdf==1.7.2
Pillow==10.2.0

# ============ Redis缓存 ============
redis==5.0.1
hiredis==2.3.2

# ============ 硬件通信 ============
pyserial==3.5
pyusb==1.2.1

# ============ 监控告警 ============
prometheus-client==0.19.0
pydantic==2.5.3
python-dotenv==1.0.0

# ============ 工具库 ============
python-dotenv==1.0.0
pydantic==2.5.3
